+++ src/components/TheFooter.vue
@@ -3,5 +3,5 @@ import { version } from '../../package.json'
 </script>
 
 <template>
-  <p class="text-center p-small">{{ version }}</p>
+  <p class="text-center p-small text-gray-600 mt-6">root.reglesdepoche {{ version }}</p>
 </template>