+++ src/presets/Root.ts
@@ -18,6 +18,37 @@ export default definePreset(Nora, {
     },
   },
   components: {
+    accordion: {
+      header: {
+        background: '{surface.950}',
+        color: '{surface.0}',
+        activeBackground: '{surface.950}',
+        activeColor: '{surface.0}',
+        hoverBackground: '{surface.950}',
+        hoverColor: '{surface.0}',
+        activeHoverBackground: '{surface.950}',
+        activeHoverColor: '{surface.0}',
+        borderColor: '{surface.950}',
+        borderWidth: '3px',
+        toggleIcon: {
+          color: '{surface.0}',
+          activeColor: '{surface.0}',
+          hoverColor: '{surface.0}',
+          activeHoverColor: '{surface.0}',
+        },
+      },
+      content: {
+        background: 'transparent',
+        borderColor: '{surface.950}',
+        borderWidth: '3px',
+      },
+    },
+    checkbox: {
+      root: {
+        background: '{amber.600}',
+        borderColor: '{amber.600}',
+      },
+    },
     datatable: {
       headerCell: {
         background: 'transparent',